Split extension start into subcommands
